Uploading for the first time

+++ src/main/java/com/blogapp/blogapp/service/impl/PostServiceImpl.java
@@ -4,6 +4,7 @@ import com.blogapp.blogapp.dataconverter.PostConverter;
 import com.blogapp.blogapp.dto.PostDto;
 import com.blogapp.blogapp.entity.Category;
 import com.blogapp.blogapp.entity.Comment;
+import com.blogapp.blogapp.entity.Person;
 import com.blogapp.blogapp.entity.Post;
 import com.blogapp.blogapp.exceptions.ResourceNotFoundException;
 import com.blogapp.blogapp.repository.CategoryRepository;
@@ -12,8 +13,11 @@ import com.blogapp.blogapp.repository.LikeRepository;
 import com.blogapp.blogapp.repository.PostRepository;
 import com.blogapp.blogapp.service.PostService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -33,58 +37,76 @@ public class PostServiceImpl implements PostService {
     private LikeRepository likeRepository;
 
     @Override
-    public PostDto createNewPost(Long categoryId, PostDto postDto) {
-        Post post = postConverter.convertDtoToEntity(postDto);
-        Category category = categoryRepository.findById(categoryId).orElseThrow(()->new ResourceNotFoundException("Category","id",categoryId));
-        post.setCategory(category);
-        Post newPost = postRepository.save(post);
+    public ResponseEntity<PostDto> createNewPost(Long categoryId, PostDto postDto, HttpSession session) {
+        Person person = (Person) session.getAttribute("person");
+        if(person.getRole().equals("ADMIN")){
+            Post post = postConverter.convertDtoToEntity(postDto);
+            Category category = categoryRepository.findById(categoryId).orElseThrow(()->new ResourceNotFoundException("Category","id",categoryId));
+            post.setCategory(category);
+            Post newPost = postRepository.save(post);
+            return new ResponseEntity<>(postConverter.convertEntityToDto(newPost), HttpStatus.CREATED);
+        }
 
-       PostDto response = postConverter.convertEntityToDto(newPost);
-        return response;
+        return null;
     }
 
     @Override
-    public List<PostDto> getAllPost() {
+    public ResponseEntity<List<PostDto>> getAllPost(HttpSession session) {
+        Person person = (Person) session.getAttribute("person");
+        if(person.getRole().equals("ADMIN")){
+            List<Post> postList = postRepository.findAll();
+            return new ResponseEntity<>(postList.stream().map(post -> postConverter.convertEntityToDto(post)).collect(Collectors.toList()),HttpStatus.OK);
+        }
 
-        List<Post> postList = postRepository.findAll();
-        return postList.stream().map(post -> postConverter.convertEntityToDto(post)).collect(Collectors.toList());
+        return null;
     }
 
     @Override
-    public PostDto getPostById(Long id) {
-        Post post = postRepository.findById(id).orElseThrow(()->new ResourceNotFoundException("Post","id",id));
-        PostDto postDto1 = postConverter.convertEntityToDto(post);
+    public ResponseEntity<PostDto >getPostById(Long id,HttpSession session) {
+        Person person = (Person) session.getAttribute("person");
+        if(person.getRole().equals("ADMIN")){
+            Post post = postRepository.findById(id).orElseThrow(()->new ResourceNotFoundException("Post","id",id));
+            PostDto postDto1 = postConverter.convertEntityToDto(post);
 
-        postDto1.setLikes(likeRepository.findAllByPostId(postDto1.getId()).size());
-        postDto1.setCommentSet(commentRepository.findByPostId(postDto1.getId()));
-        return postDto1;
+            postDto1.setLikes(likeRepository.findAllByPostId(postDto1.getId()).size());
+            postDto1.setCommentSet(commentRepository.findByPostId(postDto1.getId()));
+            return new ResponseEntity<>(postDto1,HttpStatus.OK);
+        }
+        return null;
     }
 
     @Override
-    public PostDto updatePost(PostDto postDto, Long id) {
-        Post post = postRepository.findById(id).orElseThrow(()->new ResourceNotFoundException("Product","id",id));
-        post.setName(postDto.getName());
-        post.setDescription(postDto.getDescription());
-        post.setPrice(postDto.getPrice());
+    public ResponseEntity<PostDto> updatePost(PostDto postDto, Long id, HttpSession session) {
+        Person person = (Person) session.getAttribute("person");
+        if(person.getRole().equals("ADMIN")){
+            Post post = postRepository.findById(id).orElseThrow(()->new ResourceNotFoundException("Product","id",id));
+            post.setName(postDto.getName());
+            post.setDescription(postDto.getDescription());
+            post.setPrice(postDto.getPrice());
 
-        Post updatedPost = postRepository.save(post);
-        PostDto postDto1 = postConverter.convertEntityToDto(updatedPost);
-        return postDto1;
+            Post updatedPost = postRepository.save(post);
+            return new ResponseEntity<>(postConverter.convertEntityToDto(updatedPost),HttpStatus.OK);
+        }
 
+        return null;
     }
 
     @Override
-    public void deletePost(Long id) {
-        Post post = postRepository.findById(id).orElseThrow(()->new ResourceNotFoundException("Post","id",id));
-        deleteCommentsOfAPost(id);
-        postRepository.delete(post);
+    public ResponseEntity<String> deletePost(Long id,HttpSession session) {
+        Person person = (Person) session.getAttribute("person");
+        if(person.getRole().equals("ADMIN")){
+            Post post = postRepository.findById(id).orElseThrow(()->new ResourceNotFoundException("Post","id",id));
+            deleteCommentsOfAPost(id);
+            postRepository.delete(post);
+            return new ResponseEntity<>("Post deleted successfully",HttpStatus.OK);
+        }
 
+        return new ResponseEntity<>("Access Denied",HttpStatus.OK);
 
     }
 
     @Override
     public Post getPostsById(Long id) {
-
         return postRepository.findById(id).get();
     }
